add_items_check_and_delete
