+++ app/src/main/res/layout/activity_additems.xml
@@ -1,46 +1,42 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:padding="16dp"
-
-
-    >
+<?xml version="1.0" encoding="UTF-8" ?>
 <LinearLayout
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    >
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_margin="16dp"
+    android:orientation="vertical"
+    android:layout_height="match_parent">
 
     <TextView
         android:id="@+id/title"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="To-Do List"
-        android:textSize="20sp"
-        android:textStyle="bold"
-        />
-</LinearLayout>
+        android:text="@string/todo_list"
+        android:textSize="25sp"
+        android:textStyle="bold"/>
 
     <EditText
-        android:id="@+id/inputBox"
+        android:id="@+id/editText"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:hint="Enter task"
+        android:layout_marginTop="8dp"
+        android:hint="@string/enter_your_items"/>
 
-        />
 
     <Button
         android:id="@+id/addButton"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Add"/>
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recyclerView"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
+        android:layout_marginTop="8dp"
+        android:text="@string/add"
+        android:textStyle="bold"
+        android:textSize="20sp"
         />
 
 
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/recyclerView"
+        android:layout_marginTop="18dp"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:layout_weight="1"/>
 </LinearLayout>
