+++ app/src/main/res/layout/item_task.xml
@@ -13,7 +13,7 @@
 
     <TextView
         android:id="@+id/taskText"
-        android:layout_width="wrap_content"
+        android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginLeft="10dp"
         android:text="Task"
