+++ app/src/main/res/menu/nav_menu.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <group android:checkableBehavior="single">
+        <item
+            android:id="@+id/nav_view_projects"
+            android:title="View Projects"
+            android:icon="@android:drawable/ic_menu_agenda" />
+        <item
+            android:id="@+id/nav_add_project"
+            android:title="Add Project"
+            android:icon="@android:drawable/ic_input_add" />
+    </group>
+</menu>
